[#1] Rename Container to Envelope
